Modify deployment files and launch with fly
Updated .env.op to use real environment variables and modified fly.toml
with the real karmona machine and fly deployment.

Modified Dockerfile to include 2 additional lines for RESEND. These
may be obsolete, but are included for reference since they are used in
Mujer Automotriz.

The deployment was successful, but it's deployed using fly deploy
 --build-arg OP_SERVICE_ACCOUNT_TOKEN="my-service-account-token"

Since usual secrets and environment variables are not available during
build time. We must find a solution to this for Github Actions to work.