Fix deployment workflow to use correct secrets
They have been stored in our secrets and should be changed later
once the service account is set up with those extra secrets.

The current fly api key in the service account does not match with our
current app. I generated a new one from our fly account and added it.

We should discuss if we want to leave these commits for reference, or
rebase. If this doesn't work, we'll try again.